+++ skoolkit/skool2asm.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2008-2014 Richard Dymond (rjdymond@gmail.com)
+# Copyright 2008-2015 Richard Dymond (rjdymond@gmail.com)
 #
 # This file is part of SkoolKit.
 #
@@ -23,72 +23,64 @@ from . import info, get_class, VERSION, show_package_dir
 from .skoolparser import SkoolParser, CASE_LOWER, CASE_UPPER, BASE_10, BASE_16
 from .skoolasm import AsmWriter
 
-quiet = False
-
 DEF_INSTR_WIDTH = 23
 
 TAB = 'tab'
 CRLF = 'crlf'
 INSTRUCTION_WIDTH = 'instruction-width'
 
-def notify(notice):
-    if not quiet:
-        info(notice)
-
-def clock(prefix, operation, *args, **kwargs):
+def clock(quiet, prefix, operation, *args, **kwargs):
     go = time.time()
     result = operation(*args, **kwargs)
     stop = time.time()
-    notify('{0} ({1:0.2f}s)'.format(prefix, stop - go))
+    if not quiet:
+        info('{} ({:0.2f}s)'.format(prefix, stop - go))
     return result
 
-def run(skoolfile, options, parser_mode, writer_mode):
-    global quiet
-    quiet = options.quiet
-
+def run(skoolfile, options):
     # Create the parser
-    case, base, asm_mode, warn, fix_mode, create_labels = parser_mode
     if skoolfile == '-':
         fname = 'stdin'
     else:
         fname = skoolfile
-    parser = clock('Parsed {}'.format(fname), SkoolParser, skoolfile, case=case, base=base, asm_mode=asm_mode, warnings=warn, fix_mode=fix_mode, create_labels=create_labels)
+    parser = clock(options.quiet, 'Parsed {}'.format(fname), SkoolParser, skoolfile,
+                   options.case, options.base, options.asm_mode, options.warn, options.fix_mode,
+                   False, options.create_labels, True, options.start, options.end)
 
-    properties = {}
+    # Write the ASM file
+    cls_name = options.writer or parser.asm_writer_class
+    if cls_name:
+        asm_writer_class = get_class(cls_name)
+        if not options.quiet:
+            info('Using ASM writer {0}'.format(cls_name))
+    else:
+        asm_writer_class = AsmWriter
+    properties = dict(parser.properties)
     if options.crlf:
         properties[CRLF] = '1'
     if options.inst_width is not None:
         properties[INSTRUCTION_WIDTH] = options.inst_width
     if options.tabs:
         properties[TAB] = '1'
-    writer_properties = {}
-    writer_properties.update(parser.properties)
-    writer_properties.update(properties)
-    crlf = writer_properties.get(CRLF, '0') != '0'
-    tab = writer_properties.get(TAB, '0') != '0'
-    lower, show_warnings = writer_mode
-    if show_warnings:
-        show_warnings = writer_properties.get('warnings', '1') != '0'
+    crlf = properties.get(CRLF, '0') != '0'
+    tab = properties.get(TAB, '0') != '0'
+    lower = options.case == CASE_LOWER
     try:
-        instr_width = int(writer_properties.get(INSTRUCTION_WIDTH, DEF_INSTR_WIDTH))
+        instr_width = int(properties.get(INSTRUCTION_WIDTH, DEF_INSTR_WIDTH))
     except ValueError:
         instr_width = DEF_INSTR_WIDTH
-
-    # Write the ASM file
-    cls_name = options.writer or parser.asm_writer_class
-    if cls_name:
-        asm_writer_class = get_class(cls_name)
-        notify('Using ASM writer {0}'.format(cls_name))
+    if options.warn:
+        show_warnings = properties.get('warnings', '1') != '0'
     else:
-        asm_writer_class = AsmWriter
-    skool_writer = asm_writer_class(parser, crlf, tab, writer_properties, lower, instr_width, show_warnings)
-    clock('Wrote ASM to stdout', skool_writer.write)
+        show_warnings = False
+    asm_writer = asm_writer_class(parser, crlf, tab, properties, lower, instr_width, show_warnings)
+    clock(options.quiet, 'Wrote ASM to stdout', asm_writer.write)
 
 def main(args):
     parser = argparse.ArgumentParser(
-        usage='skool2asm.py [options] file',
-        description="Convert a skool file into an ASM file, written to standard output. "
-                    "FILE may be\na regular file, or '-' for standard input.",
+        usage='skool2asm.py [options] FILE',
+        description="Convert a skool file into an ASM file and write it to standard output. "
+                    "FILE may\nbe a regular file, or '-' for standard input.",
         formatter_class=argparse.RawTextHelpFormatter,
         add_help=False
     )
@@ -100,6 +92,8 @@ def main(args):
                        help="Use CR+LF to end lines")
     group.add_argument('-D', '--decimal', dest='base', action='store_const', const=BASE_10,
                        help="Write the disassembly in decimal")
+    group.add_argument('-E', '--end', dest='end', metavar='ADDR', type=int, default=65536,
+                       help="Stop converting at this address")
     group.add_argument('-f', '--fixes', dest='fix_mode', metavar='N', type=int, choices=range(4), default=0,
                        help="Apply fixes:\n"
                             "  N=0: None (default)\n"
@@ -117,9 +111,11 @@ def main(args):
     group.add_argument('-q', '--quiet', dest='quiet', action='store_true',
                        help="Be quiet")
     group.add_argument('-r', '--rsub', dest='asm_mode', action='store_const', const=3, default=1,
-                       help="Use relocatability substitutions too (@rsub) (implies\n'-f 1')")
+                       help="Apply safe substitutions (@ssub) and relocatability\nsubstitutions (@rsub) (implies '-f 1')")
     group.add_argument('-s', '--ssub', dest='asm_mode', action='store_const', const=2, default=1,
-                       help="Use safe substitutions (@ssub)")
+                       help="Apply safe substitutions (@ssub)")
+    group.add_argument('-S', '--start', dest='start', metavar='ADDR', type=int, default=0,
+                       help="Start converting at this address")
     group.add_argument('-t', '--tabs', dest='tabs', action='store_true',
                        help="Use tab to indent instructions (default indentation is\n2 spaces)")
     group.add_argument('-u', '--upper', dest='case', action='store_const', const=CASE_UPPER,
@@ -136,6 +132,4 @@ def main(args):
         show_package_dir()
     if unknown_args or namespace.skoolfile is None:
         parser.exit(2, parser.format_help())
-    parser_mode = (namespace.case, namespace.base, namespace.asm_mode, namespace.warn, namespace.fix_mode, namespace.create_labels)
-    writer_mode = (namespace.case == CASE_LOWER, namespace.warn)
-    run(namespace.skoolfile, namespace, parser_mode, writer_mode)
+    run(namespace.skoolfile, namespace)